# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "lending-bank-gateway",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/bank-funds/collect-from-users": {
      "post": {
        "tags": [
          "bank-funds"
        ],
        "summary": "Collect From Users",
        "operationId": "collect_from_users_api_v1_bank_funds_collect_from_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CollectRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitAckEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bank-funds/distribute-to-users": {
      "post": {
        "tags": [
          "bank-funds"
        ],
        "summary": "Distribute To Users",
        "operationId": "distribute_to_users_api_v1_bank_funds_distribute_to_users_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DistributeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitAckEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bank-funds/refunds": {
      "post": {
        "tags": [
          "bank-funds"
        ],
        "summary": "Refund To User",
        "description": "退款北向端点（S5.6 拍板 2026-07-14：只补 refund，freeze/unfreeze 不做）。\n\n清算资金链口径（liquidation 调研）只用 collect + refund；退款经 gateway 落\nbank_txn_order（biz_type=RFND）保台账/幂等/对账覆盖，不允许上游直调 wedap 绕网关。",
        "operationId": "refund_to_user_api_v1_bank_funds_refunds_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefundRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitAckEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bank-funds/reversals": {
      "post": {
        "tags": [
          "bank-funds"
        ],
        "summary": "Reverse Transaction",
        "description": "通用冲正北向端点（对接 wedap Public.md §4.4.2，全额冲正原归集单）。\n\n落 RVSL 单（biz_type=RVSL）→ 同步调 wedap 通用冲正 → HTTP 200 即指令受理成功\n（RVSL 单 SUCCEEDED），同一事务把本地原单 SUCCEEDED→REVERSED（查不到不拦）。\n通用冲正同步无回调，不走既有 callback 摄取链。仅全额冲正，部分退款走 /refunds。",
        "operationId": "reverse_transaction_api_v1_bank_funds_reversals_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReversalRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitAckEnvelope"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/bank-funds/status": {
      "get": {
        "tags": [
          "bank-funds"
        ],
        "summary": "Query Status",
        "description": "查询本地 order 状态 + wedap 实时状态（wedap 不可用时降级为 unavailable=True）。",
        "operationId": "query_status_api_v1_bank_funds_status_get",
        "parameters": [
          {
            "name": "bizSeqNo",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bizseqno"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Query Status Api V1 Bank Funds Status Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "CollectRequest": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "currencyCode": {
            "type": "string",
            "title": "Currencycode"
          },
          "transType": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Transtype"
          },
          "txnAmount": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Txnamount",
            "description": "归集金额（DECIMAL(21,4) 字符串）。**业务上必填**；缺失/空串返 400 GW_400_VALIDATION。schema 标 optional 仅为统一错误封套，勿据此省略。"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "bizSeqNo",
          "currencyCode",
          "transType"
        ],
        "title": "CollectRequest"
      },
...
      "DistributeRequest": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "currencyCode": {
            "type": "string",
            "title": "Currencycode"
          },
          "transType": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Transtype"
          },
          "recipients": {
            "anyOf": [
              {
                "items": {
                  "additionalProperties": true,
                  "type": "object"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Recipients"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "bizSeqNo",
          "currencyCode",
          "transType"
        ],
        "title": "DistributeRequest"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MoneyWriteOutcome": {
        "type": "string",
        "enum": [
          "NOT_APPLIED",
          "PENDING",
          "UNKNOWN",
          "ACCEPTED"
        ],
        "title": "MoneyWriteOutcome",
        "description": "v2.2 §8.2 ``outcome`` 封闭枚举：同步成功之外的业务意图事实。"
      },
...
      "RefundRequest": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "currencyCode": {
            "type": "string",
            "title": "Currencycode"
          },
          "transType": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Transtype"
          },
          "refundAmount": {
            "type": "string",
            "title": "Refundamount"
          },
          "oriBizSeqNo": {
            "type": "string",
            "title": "Oribizseqno"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "bizSeqNo",
          "currencyCode",
          "transType",
          "refundAmount",
          "oriBizSeqNo"
        ],
        "title": "RefundRequest"
      },
...
      "ReversalRequest": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "currencyCode": {
            "type": "string",
            "title": "Currencycode"
          },
          "transType": {
            "type": "string",
            "maxLength": 32,
            "minLength": 1,
            "title": "Transtype"
          },
          "oriBizSeqNo": {
            "type": "string",
            "title": "Oribizseqno"
          },
          "oriTxnAmount": {
            "type": "string",
            "title": "Oritxnamount"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "bizSeqNo",
          "currencyCode",
          "transType",
          "oriBizSeqNo",
          "oriTxnAmount"
        ],
        "title": "ReversalRequest"
      },
      "SubmitAck": {
        "properties": {
          "bizSeqNo": {
            "type": "string",
            "title": "Bizseqno"
          },
          "txnStatus": {
            "type": "string",
            "title": "Txnstatus"
          },
          "orderStatus": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orderstatus"
          },
          "errorCode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errorcode"
          },
          "errorMsg": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Errormsg"
          },
          "inFlight": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inflight"
          },
          "outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyWriteOutcome"
              },
              {
                "type": "null"
              }
            ],
            "description": "业务意图事实；**同步完成的成功响应省略本字段**（由成功模型表达）。PENDING/UNKNOWN/ACCEPTED 一律不得重提，须查 statusUrl 或进对账。"
          },
          "operationStatus": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyWriteOperationStatus"
              },
              {
                "type": "null"
              }
            ]
          },
          "retryPolicy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/MoneyWriteRetryPolicy"
              },
              {
                "type": "null"
              }
            ],
            "description": "写客户端唯一可执行的重试策略。POLL_STATUS=查 statusUrl；CORRECT_AND_NEW_INTENT=修正请求并换新 bizSeqNo（bizSeqNo 即幂等键，同键重发只会拿回同一条冻结响应）；NEVER=不得重试。"
          },
          "resubmitAllowed": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resubmitallowed",
            "description": "是否允许用**同一幂等键**重提。本网关无 9000 在册的权威放行策略，故恒为 false（fail-closed）。"
          },
          "operationId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Operationid",
            "description": "durable operation 标识（= bizSeqNo）。与 statusUrl 成对出现。"
          },
          "statusUrl": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Statusurl",
            "description": "本 operation 的查单地址，受同 tenant/principal 的对象级授权。"
          }
        },
        "type": "object",
        "required": [
          "bizSeqNo",
          "txnStatus"
        ],
        "title": "SubmitAck",
        "description": "写原语提交响应 data 段最小字段契约（collect/distribute/refund/reversal 共用）。\n\n- ``txnStatus``：wedap 同步语义（SUCCESS 已同步终态 / PROCESSING 异步在途 /\n  FAILED / RESULT_UNKNOWN）\n- ``orderStatus``：gateway 台账 :class:`~app.domain.states.OrderStatus`\n  （ACCEPTED/SUBMITTED/SUCCEEDED/FAILED/RESULT_UNKNOWN/...），受理 vs 落账\n  以此字段区分；权威终态仍以 ``GET /bank-funds/status`` 为准。\n  缺失的两种情形：契约上线前受理单的幂等重放（first_response 冻结不回填）；\n  in-flight 重放（``inFlight=true``，零查询路径不读 order 行）——均应查单。\n- 未设置字段序列化省略（端点挂 ``response_model_exclude_unset``）：线格式对既有\n  调用方纯增量（不出现 ``errorCode: null`` 噪声）。不能用 model_serializer 丢 None\n  ——wrap serializer 的 dict 返回注解会把 openapi schema 打回 additionalProperties。"
      },
      "SubmitAckEnvelope": {
        "properties": {
          "success": {
            "type": "boolean",
            "title": "Success"
          },
          "data": {
            "$ref": "#/components/schemas/SubmitAck"
          },
          "error": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "trace_id": {
            "type": "string",
            "title": "Trace Id"
          }
        },
        "type": "object",
        "required": [
          "success",
          "data",
          "error",
          "trace_id"
        ],
        "title": "SubmitAckEnvelope",
        "description": "写原语提交 200 响应统一 envelope（:func:`app.core.envelope.ok` 的类型化形态）。"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}